tests-simulate-prod-data/requirements.txt
click==8.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
sqlalchemy==1.4.54
psycopg2-binary==2.9.11
